requirements.txt
PyPDF2
faiss-cpu
transformers==4.41.2
torch==2.3.1
accelerate==0.31.0
nltk
pandas
numpy
streamlit
python-dotenv
groq
